Fixed compiling errors with -Wextra

+++ maos/mvm_client.c
@@ -42,10 +42,10 @@ static int sock_mvm=-1;
 
 #define WRITE_ARR(p,n,type)						\
     {									\
-	int ndone;							\
+	size_t ndone;							\
 	if((ndone=write(sock_mvm, p, sizeof(type)*(n)))!=sizeof(type)*(n)){ \
 	    perror("write");						\
-	    error("error writing. want %ld wrote %d\n",(long)(n)*sizeof(type), ndone); \
+	    error("error writing. want %ld wrote %zu\n",(long)(n)*sizeof(type), ndone); \
 	}								\
     }
 #define READ_CMD(p) READ_ARR(p,N_CMD,int)
